chore(cb2-0000): add view and summary mode
